add detection and update svm

+++ models/knn/EuclideanDistance.m
@@ -1,3 +1,3 @@
 function dEuc = EuclideanDistance(sample1, sample2)
-    dEuc = sqrt(sum((sample1 - sample2).^ 2, 2));
+    dEuc = sqrt(sum((double(sample1) - double(sample2)).^ 2, 2));
 end
